tests/wpt/tests/css/css-anchor-position/position-fallback-dynamic.html

<!DOCTYPE html>
<title>CSS Anchor Positioning Test: Dynamically change position via position-fallback property</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#fallback">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  body { margin: 0; }

  @position-fallback --fallback1 {
    @try {
      left: anchor(--a1 right);
    }
  }
  #anchor {
    anchor-name: --a1;
    width: 100px;
    height: 100px;
  }
  #anchored {
    position: absolute;
    width: 100px;
    height: 100px;
  }
</style>
<div id="anchor"></div>
<div id="anchored"></div>
<script>
  test(() => {
    assert_equals(anchored.offsetLeft, 0);
  }, "Initial static left position is 0");

  test(() => {
    anchored.style.positionFallback = "--fallback1";
    assert_equals(anchored.offsetLeft, 100);
  }, "Left position set to right edge of anchor with @position-fallback");
</script>
